# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FluxPay",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/freeze-agent": {
      "post": {
        "tags": [
          "Admin Emergency Controls"
        ],
        "summary": "Freeze Agent",
        "description": "Instantly suspend an autonomous agent's credentials and prevent new payments.",
        "operationId": "freeze_agent_admin_freeze_agent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FreezeAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Freeze Agent Admin Freeze Agent Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/halt-withdrawals": {
      "post": {
        "tags": [
          "Admin Emergency Controls"
        ],
        "summary": "Halt Withdrawals",
        "description": "Toggle circuit breaker halting all outbound treasury and agent withdrawals.",
        "operationId": "halt_withdrawals_admin_halt_withdrawals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CircuitToggleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Halt Withdrawals Admin Halt Withdrawals Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/emergency-stop": {
      "post": {
        "tags": [
          "Admin Emergency Controls"
        ],
        "summary": "Emergency Stop",
        "description": "Platform-wide emergency stop. Rejects all payment ingress immediately.",
        "operationId": "emergency_stop_admin_emergency_stop_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CircuitToggleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Emergency Stop Admin Emergency Stop Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CircuitToggleRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "reason": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Reason"
          },
          "operator_id": {
            "type": "string",
            "minLength": 1,
            "title": "Operator Id",
            "default": "OPERATOR"
          }
        },
        "type": "object",
        "required": [
          "enabled",
          "reason"
        ],
        "title": "CircuitToggleRequest",
        "description": "Payload to toggle platform-wide circuit breakers."
      },
...
      "FreezeAgentRequest": {
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "reason": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "reason"
        ],
        "title": "FreezeAgentRequest",
        "description": "Payload to suspend an autonomous agent."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}